- remove deprecated FutureHelper API - reduce Vert.x dependency by moving Timers to a Java Timer implementation
We are having some problems with Netty (the event loop Vert.x is based
on) waking too often and that could be due to periodic jobs. Regardless,
the non-Vert.xed API is cleaner and simpler and allows us to easily
account for schedule drifts when using LocalTime for the schedule.