fromAsync: show original call stack trace when propagating exceptions
Because of the nature of Java's "exceptional completion" encoding, the `CompletionException` wrapper propagated to the catch site contains the Java internal completion mechanism stack trace.

Instead, we can pay a bit of CPU during each call and take a snapshot of the original call site, then in the case of an error - inject it into the `CompletionException` wrapper. This will allow much much easier debugging of async exceptions.